Core Tutorial/CMake: fix how files are configured and installed
configure_file was used to install a file during the invocation of "cmake",
which is wrong. Instead, it must be used to locally transform a file from
source to build dir.

Also, since the Parameter Framework's v2.3.0 release, the "Folder" attribute of
the PluginLocation tag can be empty. It is therefore useless to configure it.

Signed-off-by: David Wagner <david.wagner@intel.com>